Implementation of the separation
Instead of forcing a DB who wants to run the workload to contribute to
the codebase, make this a library, exposing the DB trait.

This way, they just need to include the library, implement the DB trait
for their database and pass things on to ycsb_run.

This is a first stab at it, it can still be upgraded.
