+++ Cargo.toml
@@ -1,18 +1,9 @@
-[package]
-name = "ycsb"
-version = "0.0.0"
-authors = ["Pekka Enberg <penberg@chiselstrike.com>"]
-edition = "2021"
+[workspace]
 
-[dependencies]
-anyhow = "1.0"
-itertools = "0.10"
-log = "0.4"
-rand = {version = "0.8", features = ["small_rng"]}
-rusqlite = "0.28"
-serde = { version = "1.0.130", features = ["derive"] }
-structopt = "0.3.23"
-toml = "0.5.8"
+members = [
+    "lib",
+    "examples/sqlite",
+]
 
 [profile.release]
 debug = true
