+++ examples/sqlite/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "sqlite"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+itertools = "0.10"
+rusqlite = "0.28"
+structopt = "0.3.23"
+toml = "0.5.8"
+ycsb = { "path" = "../../lib" }
