+++ lib/src/generator.rs
