+++ lib/src/generator/acknowledged_counter_generator.rs
