+++ lib/src/generator/constant_generator.rs
