+++ lib/src/generator/counter_generator.rs
