+++ lib/src/generator/discrete_generator.rs
