+++ lib/src/properties.rs
